Re-run .\create.sh for Database, Group Owner has been added to maintain Delete(Done) and Edit(In Progress), You can now join group by Invitation Code, You can leave the group, You can not join the same group with same key multiple times on same user. //TODO Show Lists to group members

+++ java/src/main/java/com/techelevator/dao/JdbcGroupDao.java
@@ -27,15 +27,16 @@ public class JdbcGroupDao implements GroupDao {
         //System.out.println(getGroupByName(group.getGroupName()));
 
 
-        String sql = "INSERT INTO groups (group_name) VALUES (?); ";
+        String sql = "INSERT INTO groups (group_name, group_key,owner_id) VALUES (?,?,?); ";
 
-        jdbcTemplate.update(sql, group.getGroupName());
+        jdbcTemplate.update(sql, group.getGroupName(), group.getGroupKey(), ownerId);
 
         Long groupId = getGroupByName(group.getGroupName()).getGroupId();
 
         String sql2 = "INSERT INTO group_user (group_id, user_id) VALUES (?,?);";
         return jdbcTemplate.update(sql2, groupId, ownerId) > 0;
     }
+
 
     @Override
     public boolean updateGroup(Group group) {
@@ -44,9 +45,37 @@ public class JdbcGroupDao implements GroupDao {
     }
 
     @Override
-    public void deleteGroup(Long groupId) {
-        String sql = "DELETE FROM groups WHERE group_id = ?";
-        jdbcTemplate.update(sql, groupId);
+    public void deleteGroup(Long groupId,Principal principal) {
+        UserDao userDao = new JdbcUserDao(jdbcTemplate);
+        String ownerName = principal.getName();
+        Long ownerId = (long) userDao.findIdByUsername(ownerName);
+
+        String sql = "DELETE FROM groups WHERE group_id = ? AND owner_id = ?";
+        jdbcTemplate.update(sql, groupId,ownerId);
+    }
+
+    @Override
+    public boolean joinGroup(String groupKey, Principal principal) {
+        UserDao userDao = new JdbcUserDao(jdbcTemplate);
+        String userName = principal.getName();
+        Long userId = (long) userDao.findIdByUsername(userName);
+
+        String sql = "SELECT * FROM groups WHERE group_key = ?; ";
+        String sql2 = "SELECT group_id FROM group_user WHERE group_id = ? AND user_id = ?; ";
+        String sql3 = "INSERT INTO group_user (group_id, user_id) VALUES (?,?); ";
+        SqlRowSet results = jdbcTemplate.queryForRowSet(sql, groupKey);
+
+        if (results.next()) {
+            //Check if there is a match in group_user
+            SqlRowSet results2 = jdbcTemplate.queryForRowSet(sql2, results.getLong("group_id"), userId);
+            if (results2.next()) {
+                return false;
+            } else {
+                return jdbcTemplate.update(sql3, results.getLong("group_id"), userId) > 0;
+            }
+        } else {
+            throw new IllegalArgumentException("No Match Found!");
+        }
     }
 
     @Override
@@ -107,6 +136,7 @@ public class JdbcGroupDao implements GroupDao {
         Group group = new Group();
         group.setGroupId(results.getLong("group_id"));
         group.setGroupName(results.getString("group_name"));
+        group.setGroupKey(results.getString("group_key"));
         return group;
     }
 }
